# Vectorize `find_end `