--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3719,7 +3719,7 @@ namespace {
                 // TRANSITION, DevCom-10689455, the code below could test with _mm_cmpestrc,
                 // if it has been fused with _mm_cmpestrm.
 
-                // The very last part, for any  match needle should fit, otherwise false match
+                // The very last part, for any match needle should fit, otherwise false match
                 __m128i _Data1_last            = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Mid1));
                 const auto _Match_last         = _mm_cmpestrm(_Data2, _Size_el_2, _Data1_last, _Part_size_el, _Op);
                 const unsigned _Match_last_val = _mm_cvtsi128_si32(_Match_last);