@StephanTLavavej 2024-09-12
I've pushed a merge with `main`, resolving a trivial adjacent add/edit conflict in `vector_algorithms.cpp`. No further clang-format regen was necessary.