@StephanTLavavej 2024-10-22
Thanks! :smile_cat: I pushed some fixes, please double-check.

Final perf results on my 5950X look good. The regressions for the highly pathological cases aren't too bad, and I don't think we need to add extra implementation complexity to avoid them.

Benchmark                           |   Before |    After | Speedup | Notes
------------------------------------|---------:|---------:|--------:|:-----
`classic_find_end<std::uint8_t>/0`  |  76.8 ns |  19.1 ns | 4.02    | Small, closer to end
`classic_find_end<std::uint8_t>/1`  |  73.7 ns |  21.6 ns | 3.41    | Large, closer to end
`classic_find_end<std::uint8_t>/2`  |  1311 ns |   195 ns | 6.72    | Small, closer to begin
`classic_find_end<std::uint8_t>/3`  |  1518 ns |   282 ns | 5.38    | Large, closer to begin
`classic_find_end<std::uint8_t>/4`  |  5427 ns |  6847 ns | 0.79    | Small, evil
`classic_find_end<std::uint8_t>/5`  | 16368 ns | 19204 ns | 0.85    | Large, evil
`classic_find_end<std::uint16_t>/0` |  80.3 ns |  29.9 ns | 2.69    | Small, closer to end
`classic_find_end<std::uint16_t>/1` |  76.6 ns |  34.2 ns | 2.24    | Large, closer to end
`classic_find_end<std::uint16_t>/2` |  1318 ns |   453 ns | 2.91    | Small, closer to begin
`classic_find_end<std::uint16_t>/3` |  1539 ns |   579 ns | 2.66    | Large, closer to begin
`classic_find_end<std::uint16_t>/4` |  8954 ns | 13588 ns | 0.66    | Small, evil
`classic_find_end<std::uint16_t>/5` | 27326 ns | 24983 ns | 1.09    | Large, evil
`ranges_find_end<std::uint8_t>/0`   |  80.1 ns |  19.4 ns | 4.13    | Small, closer to end
`ranges_find_end<std::uint8_t>/1`   |  76.0 ns |  21.9 ns | 3.47    | Large, closer to end
`ranges_find_end<std::uint8_t>/2`   |  1322 ns |   196 ns | 6.74    | Small, closer to begin
`ranges_find_end<std::uint8_t>/3`   |  1557 ns |   282 ns | 5.52    | Large, closer to begin
`ranges_find_end<std::uint8_t>/4`   |  8970 ns |  6845 ns | 1.31    | Small, evil
`ranges_find_end<std::uint8_t>/5`   | 27365 ns | 19232 ns | 1.42    | Large, evil
`ranges_find_end<std::uint16_t>/0`  |  80.7 ns |  30.4 ns | 2.65    | Small, closer to end
`ranges_find_end<std::uint16_t>/1`  |  76.1 ns |  35.0 ns | 2.17    | Large, closer to end
`ranges_find_end<std::uint16_t>/2`  |  1318 ns |   442 ns | 2.98    | Small, closer to begin
`ranges_find_end<std::uint16_t>/3`  |  1538 ns |   570 ns | 2.70    | Large, closer to begin
`ranges_find_end<std::uint16_t>/4`  |  8953 ns | 13605 ns | 0.66    | Small, evil
`ranges_find_end<std::uint16_t>/5`  | 27362 ns | 25015 ns | 1.09    | Large, evil