@AlexGuteniev 2024-10-22
I observe that you dropped parens in `(a == b) ? c : d` here, but not in bitset from stream optimization. Is there a rule?