@StephanTLavavej 2024-10-22
Nah, I just wasn't consistently complaining. :joy_cat: When an expression in a conditional operator is especially long then I don't find extra parens to be as objectionable.